ensures that user is only able to delete note that the user created, otherwise can't delete the note of other user

+++ lmn/views/views_notes.py
@@ -49,7 +49,8 @@ def delete_note(request, note_pk):
     note = get_object_or_404(Note, pk=note_pk)
     
     if request.method == 'POST':
-        note.delete() # todo confirming feature if user truly wants to delete the note.
+        if request.user == note.user: # checking if the request was made by the same user who added the note
+            note.delete() # todo confirming feature if user truly wants to delete the note.
         return redirect('latest_notes')
     else:
         return render(request, 'lmn/notes/note_detail.html', {'note': note})
